Fix Klipper params being non-blocking
