+++ CYD-Klipper/src/core/data_setup.h
@@ -2,6 +2,3 @@
 
 void data_loop();
 void data_setup();
-
-void freeze_request_thread();
-void unfreeze_request_thread();
