+++ CYD-Klipper/src/core/klipper/klipper_printer_panels.cpp
@@ -3,6 +3,27 @@
 #include "../../ui/ui_utils.h"
 #include "../common/constants.h"
 #include <stdio.h>
+#include "../semaphore.h"
+
+bool send_gcode_blocking(const char *gcode, bool wait = true)
+{
+    freeze_request_thread();
+    KlipperPrinter* printer = (KlipperPrinter*)get_current_printer(); // TODO: pass by ref
+    bool result = printer->send_gcode(gcode);
+    unfreeze_request_thread();
+
+    return result;
+}
+
+bool move_printer_blocking(const char* axis, float amount, bool relative)
+{
+    freeze_request_thread();
+    KlipperPrinter* printer = (KlipperPrinter*)get_current_printer(); // TODO: pass by ref
+    bool result = printer->move_printer(axis, amount, relative);
+    unfreeze_request_thread();
+
+    return result;
+}
 
 static void set_fan_speed_text(lv_event_t * e) {
     lv_obj_t * label = lv_event_get_target(e);
@@ -14,10 +35,9 @@ static void set_fan_speed_text(lv_event_t * e) {
 static void set_fan_speed(lv_event_t * e){
     int speed = (int)lv_event_get_user_data(e);
     int actual_speed = fan_percent_to_byte(speed);
-    KlipperPrinter* printer = (KlipperPrinter*)get_current_printer(); // TODO: pass by ref
     char gcode[16];
     sprintf(gcode, "M106 S%d", actual_speed);
-    printer->send_gcode(gcode);
+    send_gcode_blocking(gcode);
 }
 
 FAN_SPEED_COLUMN(set_fan_speed, klipper_fan_speed_columns)
@@ -40,24 +60,21 @@ static void set_zoffset_text_ex(lv_event_t * e) {
 
 static void set_zoffset(lv_event_t * e){
     char* offset = (char*)lv_event_get_user_data(e);
-    KlipperPrinter* printer = (KlipperPrinter*)get_current_printer(); // TODO: pass by ref
-
     char gcode[48];
     sprintf(gcode, "SET_GCODE_OFFSET Z_ADJUST=%s MOVE=1", offset);
-    printer->send_gcode(gcode);
+    send_gcode_blocking(gcode);
 }
 
 static void set_z(lv_event_t * e){
     void* ptr = lv_event_get_user_data(e);
     float value = *(float *)(&ptr);
-    KlipperPrinter* printer = (KlipperPrinter*)get_current_printer(); // TODO: pass by ref
 
     if (value < 0) {
-        printer->send_gcode("SET_GCODE_OFFSET Z=0 MOVE=1");
+        send_gcode_blocking("SET_GCODE_OFFSET Z=0 MOVE=1");
         return;
     }
 
-    printer->move_printer("Z", value, false);
+    move_printer_blocking("Z", value, false);
 }
 
 const char* zoffsets[] = { "-0.01", "-0.025", "-0.05", "-0.2" };
@@ -80,20 +97,18 @@ static void set_speed_mult_text(lv_event_t * e){
 
 static void set_speed_mult(lv_event_t * e){
     int speed = (int)lv_event_get_user_data(e);
-    KlipperPrinter* printer = (KlipperPrinter*)get_current_printer(); // TODO: pass by ref
     char gcode[16];
     sprintf(gcode, "M220 S%d", speed);
-    printer->send_gcode(gcode);
+    send_gcode_blocking(gcode);
 }
 
 static void set_speed_mult_offset(lv_event_t * e){
     int speed = (int)lv_event_get_user_data(e);
-    KlipperPrinter* printer = (KlipperPrinter*)get_current_printer(); // TODO: pass by ref
     float result = get_current_printer_data()->speed_mult * 100 + speed;
     get_current_printer_data()->speed_mult = result / 100;
     char gcode[16];
     sprintf(gcode, "M220 S%.0f", result);
-    printer->send_gcode(gcode);
+    send_gcode_blocking(gcode);
 }
 
 const char* speed_presets[] = { "50%", "100%", "150%", "200%" };
@@ -118,20 +133,19 @@ static void set_extrude_mult_text(lv_event_t * e){
 
 static void set_extrude_mult(lv_event_t * e){
     int speed = (int)lv_event_get_user_data(e);
-    KlipperPrinter* printer = (KlipperPrinter*)get_current_printer(); // TODO: pass by ref
     char gcode[16];
     sprintf(gcode, "M221 S%d", speed);
-    printer->send_gcode(gcode);
+    send_gcode_blocking(gcode);
 }
 
 static void set_extrude_mult_offset(lv_event_t * e){
     int speed = (int)lv_event_get_user_data(e);
-    KlipperPrinter* printer = (KlipperPrinter*)get_current_printer(); // TODO: pass by ref
     float result = get_current_printer_data()->extrude_mult * 100 + speed;
     get_current_printer_data()->extrude_mult = result / 100;
     char gcode[16];
     sprintf(gcode, "M221 S%.0f", result);
-    printer->send_gcode(gcode);
+    
+    send_gcode_blocking(gcode);
 }
 
 const char* extrude_presets[] = { "95%", "100%", "105%", "110%" };
@@ -145,7 +159,7 @@ lv_button_column_t extrude_mult_columns[] = {
 };
 
 static void open_fan_speed_panel(lv_event_t * e){
-    lv_create_fullscreen_button_matrix_popup(lv_scr_act(), set_fan_speed_text, klipper_fan_speed_columns, 2);
+    lv_create_fullscreen_button_matrix_popup(lv_scr_act(), set_fan_speed_text, klipper_fan_speed_columns, 3);
 }
 
 static void open_zoffset_panel(lv_event_t * e){
