+++ CYD-Klipper/src/core/semaphore.cpp
@@ -1,0 +1,28 @@
+#include "semaphore.h"
+#include <UrlEncode.h>
+#include <esp_task_wdt.h>
+
+SemaphoreHandle_t freezeRenderThreadSemaphore, freezeRequestThreadSemaphore;
+
+void semaphore_init(){
+    freezeRenderThreadSemaphore = xSemaphoreCreateMutex();
+    freezeRequestThreadSemaphore = xSemaphoreCreateMutex();
+    xSemaphoreGive(freezeRenderThreadSemaphore);
+    xSemaphoreGive(freezeRequestThreadSemaphore);
+}
+
+void freeze_request_thread(){
+    xSemaphoreTake(freezeRequestThreadSemaphore, portMAX_DELAY);
+}
+
+void unfreeze_request_thread(){
+    xSemaphoreGive(freezeRequestThreadSemaphore);
+}
+
+void freeze_render_thread(){
+    xSemaphoreTake(freezeRenderThreadSemaphore, portMAX_DELAY);
+}
+
+void unfreeze_render_thread(){
+    xSemaphoreGive(freezeRenderThreadSemaphore);
+}
