+++ CYD-Klipper/src/ui/ota_setup.cpp
@@ -5,6 +5,7 @@
 #include "../core/data_setup.h"
 #include "../conf/global_config.h"
 #include "ota_setup.h"
+#include "../core/semaphore.h"
 
 //const char *ota_url = "https://gist.githubusercontent.com/suchmememanyskill/ece418fe199e155340de6c224a0badf2/raw/0d6762d68bc807cbecc71e40d55b76692397a7b3/update.json"; // Test url
 const char *ota_url = "https://suchmememanyskill.github.io/CYD-Klipper/OTA.json"; // Prod url
